Adding more IEX Functions
Added in additional class methods including:
-bonus issue
-analyst recommendations
-cash flow
-ceo_compensation
-collections (unfinished, continuously get blank array returned, not sure if the endpoint URL is wrong or if I'm incorrectly trying to parse it)

the following are still unfinished:
-earnings_today
-extended_hour_quote
-financials_as_reported
-chart
-fundamentals
-fundamental_valuations
-historical prices
among other functions that are finished in the IEX API

What needs to be done next:
We need to combine in the TD Ameritrade API side into this class. This is being done so that we will have ONE symbol object (i.e AAPL) that will have all the above fundamental, economic, options, price data, etc associated with it. This is a key part of the SQLAlchemy implemenation where every object is mapped to the database.